+++ spec/services/bloom_trade_client/exchange_rates/convert_spec.rb
@@ -41,12 +41,12 @@ module BloomTradeClient
             })
 
             resulting_rate = described_class.(
-              base_currency: "USD",
-              counter_currency: "PHP"
+              base_currency: "PHP",
+              counter_currency: "USD"
             )
 
-            expect(resulting_rate.rate).to eq 0.02
-            expect(1.0 / resulting_rate.rate).to eq 50.0
+            expect(resulting_rate).to be_a BloomTradeClient::ConversionResult
+            expect(resulting_rate.rate).to eq 50
           end
         end
 
@@ -68,7 +68,7 @@ module BloomTradeClient
               counter_currency: "KRW",
               reserve_currency: "PHP"
             )
-            expect(resulting_rate).to eq 2_000_000
+            expect(resulting_rate.rate).to eq 2_000_000
           end
         end
 
@@ -135,7 +135,7 @@ module BloomTradeClient
                   type: rate_type.to_s,
                 )
 
-                expect(resulting_rate).to eq rate
+                expect(resulting_rate.rate).to eq rate
               end
             end
           end