+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Fix not returning `BloomTradeClient::ConversionResult` on reverse rates
+
 ## [0.16.0] - 2018-11-08
 ### Changed
 - `BloomTradeClient::ExchangeRates::Sync` now returns `expires_at` field